+++ packages/babel-plugin/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @vanilla-extract/babel-plugin
 
+## 1.1.6
+
+### Patch Changes
+
+- Updated dependencies [[`3c9b7d9`](https://github.com/seek-oss/vanilla-extract/commit/3c9b7d9f2f7cba8635e7459c36585109b6616636)]:
+  - @vanilla-extract/integration@4.0.0
+
 ## 1.1.5
 
 ### Patch Changes